Update UPDATE.md correct wrong info. about App. "asking" if you want to activate Root patching.
Using iMac 16.2; whenever an OTA update has been run and the device unattended for some hours; there has never been a 
prompt to patch or update.   Combined with the false info. on the front page of the "Official guide" viz: "Native OTA updates
Install updates the moment they come out, with native System Preferences support, just like a supported Mac" which clearly indicates that no patching is required and if/when this is followed by no prompt after the OTA update then some will commence lengthy troubleshooting or give up (I spent 4 hours before taking a wild guess that clicking on "Patching" might fix the device - and yes I hadn't digested the more detailed guidance further into the guide - although the FAQ on this could reasonably be interpreted to indicate that patching is automatic after an OTA update i.e. since there is no mention of it earlier nor in the FAQ - which only mentions it as an "after the fact" type of comment without stating what that "fact" was.